Implement notification service
Added function to create new notifications
