+++ server/src/controllers/notification.controller.ts
@@ -1,5 +1,9 @@
 import {Request, Response} from 'express';
 import {Notifications, INotification} from '../models/notifications.model';
+import {
+  createNotification,
+  groupNotificationsByType,
+} from '../services/notification.service';
 
 export class NotificationController {
   /** Retrieve all non-resolved notifications for a user */
@@ -25,10 +29,12 @@ export class NotificationController {
     });
   }
 
+  /** Retrieve a single notification by its ID */
   async getNotification(req: Request, res: Response) {
     const {notificationId} = req.params;
     // TODO: implement error response
     const notification = await Notifications.findOne({_id: notificationId});
+
     res.json({
       success: true,
       notification,
@@ -53,24 +59,3 @@ export class NotificationController {
     }
   }
 }
-
-/**
- * Group notifications by type
- * @param notifications raw array of notifications
- * @returns object with grouped notifications
- */
-const groupNotificationsByType = (notifications: INotification[]) => {
-  // Custom reducer function to implement with array.reduce
-  // Groups array by all different types
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  const group = (acc: any, cur: any) => {
-    acc[cur.type] = acc[cur.type] || []; // Create new empty array if no type is defined
-    acc[cur.type].push(cur);
-    return acc;
-  };
-
-  // Implement array.reduce method
-  // Initialize with a null object, else first notification is skipped
-  const groupedNotifications = notifications.reduce(group, Object.create(null));
-  return groupedNotifications;
-};
