+++ server/src/services/notification.service.ts
@@ -1,0 +1,27 @@
+import {Notifications, INotification} from '../models/notifications.model';
+
+/**
+ * Group notifications by type
+ * @param notifications raw array of notifications
+ * @returns object with grouped notifications
+ */
+export const groupNotificationsByType = (notifications: INotification[]) => {
+  // Custom reducer function to implement with array.reduce
+  // Groups array by all different types
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  const group = (acc: any, cur: any) => {
+    acc[cur.type] = acc[cur.type] || []; // Create new empty array if no type is defined
+    acc[cur.type].push(cur);
+    return acc;
+  };
+
+  // Implement array.reduce method
+  // Initialize with a null object, else first notification is skipped
+  const groupedNotifications = notifications.reduce(group, Object.create(null));
+  return groupedNotifications;
+};
+
+export const createNotification = async (notification: INotification) => {
+  const doc = await Notifications.create(notification);
+  console.log(doc);
+};
